# gamify-specs/src/main/resources/api-spec.yaml
openapi: '3.0.3'
info:
  version: '0.1.0'
  title: Gamify API
  description: An API that manages application-specific events, and lets you define rules to find user scores.
security:
  #  - bearerAuth: []
  - apiKeyAuth: [ ]
paths:
  /events:
    post:
      description: post a user defined event
      operationId: addEvent
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/Event'
      responses:
        '201':
          description: created
    get:
      description: get all the events
      operationId: getEvents
      responses:
        '200':
          description: success
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'


  # TODO : This should be dropped.
  /fruits:
    post:
      description: create a fruit
      operationId: createFruit
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/Fruit'
      responses:
        '201':
          description: created
    get:
      description: get the list of all fruits
      operationId: getFruits
      responses:
        '200':
          description: success
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Fruit'
  /fruits/{id}:
    get:
      description: get one fruit, by id
      operationId: getFruit
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: success
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Fruit'
components:
  securitySchemes:
    #    bearerAuth:            # arbitrary name for the security scheme
    #      type: http
    #      scheme: bearer       # warning: it HAS TO be a lowercase b for the swagger ui version used in springfox
    #      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
  schemas:
    # An Event is specific to each user of our API.
    Event:
      type: object
      properties:
        userId:
          type: string
          nullable: false
        timestamp:
          type: string
          format: date-time
          nullable: false
        type:
          type: string
          nullable: false

    # A registration is a piece of data used to identify an application.
    Registration:
      type: object
      properties:
        appId:
          type: string
          nullable: false
        password:
          type: string
          nullable: false

    # TODO : This should be dropped.
    Fruit:
      type: object
      properties:
        kind:
          type: string
        colour:
          type: string
        size:
          type: string
        weight:
          type: string
        expirationDate:
          type: string
          format: date
          nullable: false
        expirationDateTime:
          type: string
          format: date-time